requirements.txt
bencodepy
docker
fastapi
httpx
loguru
passlib[bcrypt]
peewee
pytest
pytest-asyncio
python-dotenv
python-multipart
requests
transmission-rpc
uvicorn
